English.md
I spent 4 hours researching it, and it is a semi-finished product, as can be seen from the version name. I am a student with limited time, but I will use my free time to continuously improve and update its algorithms and UI. My energy and intelligence are limited, and I may not be able to come up with better algorithms. Please help me!